refactor(e2e): extract continueDebugSession to shared debugHelpers
