+++ packages/salesforcedx-vscode-apex-replay-debugger/test/playwright/specs/apexReplayDebugger.desktop.spec.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { expect, type Page } from '@playwright/test';
+import { expect } from '@playwright/test';
 import {
   APEX_TRACE_FLAG_STATUS_BAR,
   clearOutputChannel,
@@ -24,7 +24,6 @@ import {
   setupNetworkMonitoring,
   validateNoCriticalErrors,
   waitForOutputChannelText,
-  WORKBENCH,
   waitForQuickInputFirstOption
 } from '@salesforce/playwright-vscode-ext';
 
@@ -32,24 +31,7 @@ import apexLogNls from 'salesforcedx-vscode-apex-log/package.nls.json';
 import metadataNls from 'salesforcedx-vscode-metadata/package.nls.json';
 import packageNls from '../../../package.nls.json';
 import { test } from '../fixtures';
-
-/** Continue debug session (dismiss hover, Escape, then F5). Repeats until session ends. */
-const continueDebugSession = async (page: Page, maxContinues = 2): Promise<void> => {
-  const toolbar = page.locator('.debug-toolbar');
-  for (let i = 0; i < maxContinues; i++) {
-    await toolbar.waitFor({ state: 'visible', timeout: 15_000 });
-    // Click editor area to dismiss search-bar hover that can cover debug toolbar and block F5
-    await page.locator(`${WORKBENCH} .editor-instance .view-lines`).first().click({ force: true });
-    await page.keyboard.press('Escape');
-    await page.keyboard.press('F5');
-    const sessionEnded = await expect(toolbar)
-      .not.toBeVisible({ timeout: 30_000 })
-      .then(() => true)
-      .catch(() => false);
-    if (sessionEnded) break;
-  }
-  await expect(toolbar).not.toBeVisible({ timeout: 45_000 });
-};
+import { continueDebugSession } from '../helpers/debugHelpers';
 
 test('Apex Replay Debugger: trace flag, exec anon, replay from log and test class', async ({ page }) => {
   test.setTimeout(600_000);
