+++ packages/salesforcedx-vscode-apex-replay-debugger/test/playwright/specs/checkpoints.desktop.spec.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { expect, type Page } from '@playwright/test';
+import { expect } from '@playwright/test';
 import {
   APEX_TRACE_FLAG_STATUS_BAR,
   clearOutputChannel,
@@ -33,27 +33,11 @@ import apexLogNls from 'salesforcedx-vscode-apex-log/package.nls.json';
 import metadataNls from 'salesforcedx-vscode-metadata/package.nls.json';
 import packageNls from '../../../package.nls.json';
 import { test } from '../fixtures';
+import { continueDebugSession } from '../helpers/debugHelpers';
 
 // Localized fragment of the base adapter's `heap_dump_error_wrap_up_text` (salesforcedx-apex-replay-debugger
 // i18n) — emitted to the Debug Console only when host↔adapter heapDumpResults wiring fails.
 const HEAP_DUMP_ERROR_TEXT = /Problems were encountered while retrieving heap dump information/;
-
-/** Continue debug session (dismiss hover, Escape, then F5). Repeats until session ends. */
-const continueDebugSession = async (page: Page, maxContinues = 3): Promise<void> => {
-  const toolbar = page.locator('.debug-toolbar');
-  for (let i = 0; i < maxContinues; i++) {
-    await toolbar.waitFor({ state: 'visible', timeout: 15_000 });
-    await page.locator(`${WORKBENCH} .editor-instance .view-lines`).first().click({ force: true });
-    await page.keyboard.press('Escape');
-    await page.keyboard.press('F5');
-    const sessionEnded = await expect(toolbar)
-      .not.toBeVisible({ timeout: 30_000 })
-      .then(() => true)
-      .catch(() => false);
-    if (sessionEnded) break;
-  }
-  await expect(toolbar).not.toBeVisible({ timeout: 45_000 });
-};
 
 test('Checkpoints: Toggle Checkpoint and Update Checkpoints in Org', async ({ page }) => {
   test.setTimeout(600_000);
